fix(membrane): upgrade generation counters to u64 and rewrite persistence algorithm
Generation Counter Upgrade (u32→u64):
- Upgrade generation tracking from u32 to u64 across TLS cache, RCU,
  seqlock, and ptr_validator to support extended application lifetimes
  without counter overflow
- Update adversarial test fixtures to use u64 generation values

Persistence Homology Detector:
- Replace naive diameter computation with proper Kruskal-style minimum
  spanning forest (MSF) using union-find, enabling correct topological
  clustering for allocation pattern analysis

RCU Improvements:
- Add RcuCell::load_versioned() for atomic epoch/snapshot pairing,
  fixing potential races between version reads and data snapshots

Seqlock Improvements:
- Enhance writer acquisition with try_lock and contention metrics
  to reduce spinning under high load
- Hold guard properly during load_versioned reads

Resolver:
- Mark IPv6 HostsAddress variant with #[allow(dead_code)] pending
  future activation of IPv6 /etc/hosts resolution

+++ crates/frankenlibc-harness/tests/bd18qq4_uaf_adversarial_test.rs
@@ -71,15 +71,15 @@ impl ScenarioConfig {
 #[derive(Debug, Clone, Copy)]
 struct FreedRecord {
     addr: usize,
-    alloc_generation: u32,
-    post_free_generation: u32,
+    alloc_generation: u64,
+    post_free_generation: u64,
 }
 
 #[derive(Debug, Clone, Copy)]
 struct ProbeResult {
     addr: usize,
-    alloc_generation: u32,
-    observed_generation: Option<u32>,
+    alloc_generation: u64,
+    observed_generation: Option<u64>,
     strict_detected: bool,
     strict_can_read: bool,
     strict_can_write: bool,
